% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_ui.R
\name{create_disclaimer_text}
\alias{create_disclaimer_text}
\title{Create Disclaimer Text}
\usage{
create_disclaimer_text()
}
\value{
Shiny UI element
}
\description{
Creates the disclaimer text for the application
}
