% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_ui.R
\name{create_download_full_data_button}
\alias{create_download_full_data_button}
\title{Create Download Full Data Button}
\usage{
create_download_full_data_button(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID}
}
\value{
Shiny UI element
}
\description{
Creates a download button for the full dataset
}
