% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_ui.R
\name{create_no_results_ui}
\alias{create_no_results_ui}
\title{Create No Results UI}
\usage{
create_no_results_ui()
}
\value{
Shiny UI element
}
\description{
Creates UI for when no search results are found
}
