% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_ui.R
\name{create_results_cards}
\alias{create_results_cards}
\title{Create Results Cards}
\usage{
create_results_cards(search_results_df, ns)
}
\arguments{
\item{search_results_df}{Data frame with search results}

\item{ns}{Namespace function}
}
\value{
Shiny UI element
}
\description{
Creates cards for search results display
}
