% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_data.R
\name{estimate_tokens}
\alias{estimate_tokens}
\title{Estimate Token Count for Text}
\usage{
estimate_tokens(text)
}
\arguments{
\item{text}{A character string to analyze}
}
\value{
Integer estimate of token count
}
\description{
Approximates the number of tokens (as used by models like GPT) in a given text based on character classes.
}
