% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_data.R
\name{generate_markdown_from_metadata}
\alias{generate_markdown_from_metadata}
\title{Generate Metadata Markdown from ODS Dataset}
\usage{
generate_markdown_from_metadata(dataset_id, output_dir = ".")
}
\arguments{
\item{dataset_id}{The dataset identifier on data.tg.ch}

\item{output_dir}{The output directory for the markdown file (default = current directory)}
}
\value{
No return value. Writes `metadata.md` to disk.
}
\description{
Downloads and formats metadata for a given dataset ID from data.tg.ch into a Markdown document.
}
