% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset_view.R
\name{mod_dataset_view_server}
\alias{mod_dataset_view_server}
\title{Dataset View Module Server}
\usage{
mod_dataset_view_server(id, chat_module, selected_dataset, dataset_id)
}
\arguments{
\item{id}{Module namespace ID}

\item{chat_module}{Chat module object}

\item{selected_dataset}{Reactive value with selected dataset title}

\item{dataset_id}{Reactive value with dataset ID}
}
\description{
Server logic for displaying dataset information and handling downloads
}
