% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset_view.R
\name{mod_dataset_view_ui}
\alias{mod_dataset_view_ui}
\title{Dataset View Module UI}
\usage{
mod_dataset_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates the UI for displaying dataset information and data table
}
