% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_search.R
\name{mod_search_server}
\alias{mod_search_server}
\title{Search Module Server}
\usage{
mod_search_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
List with selected_dataset reactive
}
\description{
Server logic for the dataset search functionality
}
