% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_search.R
\name{mod_search_ui}
\alias{mod_search_ui}
\title{Search Module UI}
\usage{
mod_search_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates the UI for the dataset search functionality
}
