% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sql_query.R
\name{mod_sql_query_server}
\alias{mod_sql_query_server}
\title{SQL Query Module Server}
\usage{
mod_sql_query_server(id, chat_module)
}
\arguments{
\item{id}{Module namespace ID}

\item{chat_module}{Chat module object}
}
\description{
Server logic for SQL query execution
}
