% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sql_query.R
\name{mod_sql_query_ui}
\alias{mod_sql_query_ui}
\title{SQL Query Module UI}
\usage{
mod_sql_query_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Creates the UI for SQL query input and results
}
