% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_data.R
\name{process_search_results}
\alias{process_search_results}
\title{Process Search Results}
\usage{
process_search_results(search_results, catalog)
}
\arguments{
\item{search_results}{Raw search results}

\item{catalog}{vector of non-geodata dataset_id}
}
\value{
Processed data frame
}
\description{
Processes and filters search results
}
