% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_querychat.R
\name{querychat_sidebar2}
\alias{querychat_sidebar2}
\title{QueryChat Sidebar Wrapper}
\usage{
querychat_sidebar2(id, width = 400, height = "100\%", ...)
}
\arguments{
\item{id}{Module ID for namespace}

\item{width}{Width of the sidebar}

\item{height}{Height of the sidebar}

\item{...}{Additional arguments passed to `querychat_ui2()`}
}
\description{
Wraps the querychat UI into a collapsible sidebar
}
\keyword{internal}
