% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_querychat.R
\name{querychat_ui2}
\alias{querychat_ui2}
\title{QueryChat UI Module}
\usage{
querychat_ui2(id)
}
\arguments{
\item{id}{Module ID for namespace}
}
\description{
Renders the chat interface and applies styling
}
\keyword{internal}
