% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_data.R
\name{search_datasets}
\alias{search_datasets}
\title{Search Datasets}
\usage{
search_datasets(search_term)
}
\arguments{
\item{search_term}{Search term}
}
\value{
Search results
}
\description{
Performs vector search on the dataset catalog
}
