% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_data.R
\name{vector_search_ods}
\alias{vector_search_ods}
\title{Perform Vector Search on Thurgau ODS}
\usage{
vector_search_ods(searchterm)
}
\arguments{
\item{searchterm}{A search phrase}
}
\value{
A list of dataset metadata results
}
\description{
Executes a similarity-based vector search using the ODS API from data.tg.ch.
}
